Five miscompilations, found by pointing the oracle at the tier
test262 barely tested the compiled tier. The threshold is eight calls and
most of the suite calls a function once, so 42739/42740 with the JIT on
was largely the interpreter's score reported as the tier's.
GOANT_JIT_THRESHOLD=1 (and the OSR threshold beside it) compiles
everything and makes the interpreter the oracle it should always have
been. Four bugs came out of the first run; a fifth came out of Octane's
pdfjs, which has been failing under the JIT and passing without it for as
long as there has been a JIT.

A BRANCH ACROSS THE REGISTER WINDOW. The refill that brings an operand
back into a register is emitted after the template, so on a conditional
branch it sat on the fall-through path alone: the taken edge carried on
with the condition still in the register the window said held an operand.
It cannot be hoisted above the branch either — the slot re-entering the
window is the same register the condition is in — so the shape is
refused. It takes ten live operands around a branch to arise, which in
practice means an array literal with a conditional element past the
ninth. test262's Error.prototype.stack setter test builds exactly one,
and every label in it came out as the first character of another label.
Two functions of Octane's 6976.

THREE BIGINTS TREATED AS NUMBERS. `~`, `-`, `++`, `--` and the six
bitwise operators each declared their result a Number. That is true on
the fast side of their guard and false on the other, where the runtime
answers a BigInt for a BigInt — so `~~1n` was -1, and a local assigned
from any of them was marked numeric and then used unguarded. Each now
answers whatever its operand was, which is what the emitter already did
for `-` and what the other three and the locals analysis did not.

NEW.TARGET THROUGH A DIRECT EVAL. A compiled frame published none, so
`eval('new.target')` in a constructor answered undefined. The function
never names new.target — if it did it would not compile, that opcode
having no template — but its eval can ask anyway.

A TAIL CALLEE THAT DECLINES. Declining is compiled code saying "nothing
has happened, run me in the interpreter instead", reported by returning
false to whoever entered the frame. After a tail call that is a lie: the
frame belongs to the callee and the caller it took over has already run
to its return, so the caller's caller ran the ORIGINAL function a second
time. pdfjs read the same PDF stream twice and called the document
malformed. The callee is now run through the general path instead — one
Go frame deeper than a proper tail call, and only when a compiled callee
refuses the arguments a tail call brought it.

Finding that one is worth recording. The bisect has to be over a HASH OF
THE FUNCTION NAME (GOANT_JIT_MASK, 64 buckets) and not over compile
order, because compiling one function changes how often the others are
entered and the order is not stable between runs. That took 40,000 lines
of pdfjs down to two functions in about fifteen runs.

Alongside, four costs that profiling named:

  - numberToString ran every number through a FormatFloat and four
    string operations. Most of the numbers a program stringifies are
    small integers, which FormatInt produces directly and which the
    spec's first case covers exactly. It was 18% of raytrace.
  - dropping the compiled frame's captured-upvalue map was a hashed
    delete on every frame entry AND every way out, whether anything had
    been captured or not — 7% of earley-boyer. Indexed by depth rather
    than keyed by it, it is a bounds check and a nil store.
  - the collector asked "can this type hold a Value" once per SLICE
    ELEMENT rather than once per element type: 5% of earley-boyer in a
    sync.Map lookup.
  - publishFrame cleared a whole vmFrame for a caller that immediately
    overwrote six of its fields.

And one measured and rejected, recorded where the next person will look:
widening the property probe's tag test to the whole object family takes
raytrace's cache hit rate from 82.6% to 92.5% and removes 5.1 million
helper round trips, and does not show up in the score. The family is not
contiguous, so it costs two compares on every property read to save a
round trip on one in ten.

test262 core 42739/42740 at the default threshold AND with everything
compiled; language 23173/23173 under GC poison. Static Octane coverage
6969 of 6976. pdfjs runs.

+++ internal/engine/jit_call_amd64_test.go
@@ -31,7 +31,7 @@ func BenchmarkCallIntoATinyFunction(b *testing.B) {
 		// Warm past the tier's threshold so the measured calls are the steady
 		// state rather than the compile.
 		args := []Value{tov(1)}
-		for i := 0; i < 4*jitThreshold; i++ {
+		for i := int32(0); i < 4*jitThreshold; i++ {
 			if _, e := rt.callValue(fnVal, mkundef(), args); e != nil {
 				b.Fatal("threw")
 			}
@@ -249,7 +249,7 @@ func BenchmarkCallIntoALoop(b *testing.B) {
 			b.Fatal(err)
 		}
 		args := []Value{tov(50)}
-		for i := 0; i < 4*jitThreshold; i++ {
+		for i := int32(0); i < 4*jitThreshold; i++ {
 			if _, e := rt.callValue(fnVal, mkundef(), args); e != nil {
 				b.Fatal("threw")
 			}
